requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=0.6,!=0.7,<1.0
Django>=1.4.2,<1.8
python-roboticeclient>=0.1.1
six>=1.9.0